pyproject.toml
[build-system]
requires = ["setuptools>=61.2", "setuptools-rust"]
build-backend = "setuptools.build_meta"

[tool.mypy]
warn_redundant_casts = true
warn_unused_configs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "requirements.*",
    "openai.*",
]
ignore_missing_imports = true

[project]
name = "buildlog-consultant"
authors = [{name = "Jelmer Vernooĳ", email = "jelmer@jelmer.uk"}]
description = "buildlog parser and analyser"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python_debian",
    "PyYAML",
    "requirements-parser",
]
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/jelmer/buildlog-consultant"
Repository = "https://github.com/jelmer/buildlog-consultant.git"

[project.optional-dependencies]
chatgpt = ["openai"]
dev = ["ruff==0.6.1"]

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
where = ["py"]
include = ["buildlog_consultant*"]

[tool.setuptools.package-data]
buildlog_consultant = ["py.typed"]

[tool.setuptools.dynamic]
version = {attr = "buildlog_consultant.__version__"}

[tool.ruff.lint]
select = [
    "ANN",
    "D",
    "E",
    "F",
    "I",
    "UP",
]
ignore = [
    "ANN001",
    "ANN002",
    "ANN003",
    "ANN101",
    "ANN102",
    "ANN201",
    "ANN202",
    "ANN204",
    "ANN206",
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D107",
    "E501",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.cibuildwheel]
skip = "*-win32 *musllinux*"
before-build = "pip install -U setuptools-rust && rustup default stable && rustup show"
environment = {PATH = "$HOME/.cargo/bin:$PATH"}

[tool.cibuildwheel.linux]
before-build = "pip install -U setuptools-rust && curl https://sh.rustup.rs -sSf | sh -s -- --default-toolchain=stable --profile=minimal -y && rustup show"

[tool.cibuildwheel.macos]
before-build = "rustup target add aarch64-apple-darwin"
